Move AnsibleSnippetsProcessor to ssg/ansible.py
